remove web, ios and android folders from kitchen
