+++ src/kitchen/components/lines/Lines.stories.tsx
@@ -1,6 +1,6 @@
 import { Lines, LinesProps } from './index';
-import type { Story } from '../../../../@types/storybook-emotion-10-fixes';
-import { asPlayground, asChromaticStory } from '../../../../lib/story-intents';
+import type { Story } from '../../../@types/storybook-emotion-10-fixes';
+import { asPlayground, asChromaticStory } from '../../../lib/story-intents';
 
 export default {
 	title: 'Editorial/src-ed-lines/Lines',
