+++ src/kitchen/components/lines/dashed.ts
