+++ src/kitchen/components/lines/dotted.ts
