+++ src/kitchen/components/lines/squiggly.ts
