+++ src/kitchen/components/logo/Logo.stories.tsx
@@ -1,6 +1,6 @@
 import { Logo, LogoProps } from './Logo';
-import type { Story } from '../../../../@types/storybook-emotion-10-fixes';
-import { asPlayground, asChromaticStory } from '../../../../lib/story-intents';
+import type { Story } from '../../../@types/storybook-emotion-10-fixes';
+import { asPlayground, asChromaticStory } from '../../../lib/story-intents';
 
 export default {
 	title: 'Editorial/src-ed-logo/Logo',
