+++ src/kitchen/components/logo/package.json
