+++ src/kitchen/components/quote-icon/QuoteIcon.stories.tsx
@@ -1,7 +1,7 @@
 import { Design, Display, Pillar, Special } from '@guardian/types';
 import { QuoteIcon, QuoteIconProps } from './QuoteIcon';
-import type { Story } from '../../../../@types/storybook-emotion-10-fixes';
-import { asPlayground, asChromaticStory } from '../../../../lib/story-intents';
+import type { Story } from '../../../@types/storybook-emotion-10-fixes';
+import { asPlayground, asChromaticStory } from '../../../lib/story-intents';
 
 const defaultFormat = {
 	display: Display.Standard,
