+++ src/kitchen/components/quote-icon/QuoteIcon.tsx
