+++ src/kitchen/components/quote-icon/README.md
