+++ src/kitchen/components/quote-icon/index.tsx
