+++ src/kitchen/components/quote-icon/package.json
