+++ src/kitchen/components/starRating/README.md
