+++ src/kitchen/components/starRating/StarRating.tsx
