+++ src/kitchen/components/starRating/index.tsx
