+++ src/kitchen/components/starRating/package.json
