+++ src/kitchen/components/starRating/star.ts
