+++ src/kitchen/tsconfig.json
@@ -1,15 +1,15 @@
 {
-  "extends": "../../../../../tsconfig.json",
+  "extends": "../../tsconfig.json",
   "compilerOptions": {
     "outDir": "dist/types",
     "declaration": true,
     "emitDeclarationOnly": true
   },
   "include": ["."],
-  "exclude": ["dist", "*.stories.tsx"],
+  "exclude": ["dist"],
   "references": [
     {
-      "path": "../../../../core/helpers"
+      "path": "../core/foundations"
     }
   ]
 }
